Removed retention=SOURCE (so used retention=CLASS) because annotation used in annotations like Qualify.Auto do not work if the Auto annotation type isn't in the same project
Although this is a compile time metamodel generator, might be useful that to keep the retention=CLASS in general, because some times will be interesting to further introspect the metadata back to other models, so for example, you may want to see if an other-project-type has been annotated with Qualify to link it or not, if you want to do that you need retention=CLASS (uhm, but you can check the existence of the static metamodel, yep, right, not the best example). Anyways, annotation to be used in annotation requires retention=CLASS (like Qualify.Auto) or it only works in the same project that are defined.